Backend/requirements.txt
fastapi
python-dotenv
websockets
aiokafka
fastapi-socketio
pymongo